pyproject.toml
[project]
name = "image2svg-mcp"
version = "0.1.0"
description = "MCP server that converts raster images (PNG/JPG/WEBP) to SVG vector format"
readme = "README.md"
license = "Apache-2.0"
authors = [{ name = "botmonster", email = "btm@botmonster.com" }]
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.2.0",
    "httpx>=0.28.1",
    "pillow>=12.2.0",
    "vtracer>=0.6.15",
]

[project.urls]
Homepage = "https://botmonster.com/image2svg"
Repository = "https://github.com/botmonster/image2svg-mcp"

[project.scripts]
image2svg-mcp = "image2svg_mcp.server:main"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["pytest>=9.0.2", "pytest-asyncio>=1.3.0"]
